+++ lib/compat/wordpress-6.1/script-loader.php
@@ -65,7 +65,7 @@ function gutenberg_enqueue_stored_styles() {
 		}
 		// Chain core store ids to signify what the styles contain.
 		$style_tag_id             .= '-' . $style_key;
-		$compiled_core_stylesheet .= gutenberg_style_engine_get_stylesheet_from_store( $style_key );
+		$compiled_core_stylesheet .= gutenberg_style_engine_get_stylesheet_from_context( $style_key );
 	}
 
 	// Combine Core styles.
@@ -81,7 +81,7 @@ function gutenberg_enqueue_stored_styles() {
 		if ( in_array( $store_name, $core_styles_keys, true ) ) {
 			continue;
 		}
-		$styles = gutenberg_style_engine_get_stylesheet_from_store( $store_name );
+		$styles = gutenberg_style_engine_get_stylesheet_from_context( $store_name );
 		if ( ! empty( $styles ) ) {
 			$key = "wp-style-engine-$store_name";
 			wp_register_style( $key, false, array(), true, true );
@@ -104,7 +104,7 @@ function gutenberg_enqueue_stored_styles() {
  * @param array $nodes The nodes to filter.
  * @return array A filtered array of style nodes.
  */
-function filter_out_block_nodes( $nodes ) {
+function gutenberg_filter_out_block_nodes( $nodes ) {
 	return array_filter(
 		$nodes,
 		function( $node ) {
@@ -144,7 +144,7 @@ function gutenberg_enqueue_global_styles() {
 	 * This removes the CSS from the global-styles stylesheet and adds it to the inline CSS for each block.
 	 * This filter has to be registered before we call gutenberg_get_global_stylesheet();
 	 */
-	add_filter( 'gutenberg_get_style_nodes', 'filter_out_block_nodes', 10, 1 );
+	add_filter( 'gutenberg_theme_json_get_style_nodes', 'gutenberg_filter_out_block_nodes', 10, 1 );
 
 	$stylesheet = gutenberg_get_global_stylesheet();
 	if ( empty( $stylesheet ) ) {
@@ -156,7 +156,7 @@ function gutenberg_enqueue_global_styles() {
 	wp_enqueue_style( 'global-styles' );
 
 	// add each block as an inline css.
-	wp_add_global_styles_for_blocks();
+	gutenberg_add_global_styles_for_blocks();
 }
 
 remove_action( 'wp_enqueue_scripts', 'wp_enqueue_global_styles' );