+++ packages/babel-plugin-makepot/index.js
@@ -33,15 +33,7 @@
  */
 
 const { po } = require( 'gettext-parser' );
-const {
-	pick,
-	reduce,
-	forEach,
-	sortBy,
-	isEqual,
-	merge,
-	isEmpty,
-} = require( 'lodash' );
+const { pick, reduce, isEqual, merge, isEmpty } = require( 'lodash' );
 const { relative, sep } = require( 'path' );
 const { writeFileSync } = require( 'fs' );
 
@@ -130,7 +122,7 @@ function getExtractedComment( path, _originalNodeLine ) {
 	}
 
 	let comment;
-	forEach( node.leadingComments, ( commentNode ) => {
+	Object.values( node.leadingComments ?? {} ).forEach( ( commentNode ) => {
 		let line = 0;
 		if ( commentNode && commentNode.loc && commentNode.loc.end ) {
 			line = commentNode.loc.end.line;
@@ -193,6 +185,20 @@ function isSameTranslation( a, b ) {
 	return isEqual(
 		pick( a, VALID_TRANSLATION_KEYS ),
 		pick( b, VALID_TRANSLATION_KEYS )
+	);
+}
+
+/**
+ * Sorts multiple translation objects by their reference.
+ * The reference is where they occur, in the format `file:line`.
+ *
+ * @param {Array} translations Array of translations to sort.
+ *
+ * @return {Array} Sorted translations.
+ */
+function sortByReference( translations = [] ) {
+	return [ ...translations ].sort( ( a, b ) =>
+		a.comments.reference.localeCompare( b.comments.reference )
 	);
 }
 
@@ -323,46 +329,39 @@ module.exports = () => {
 						( memo, file ) => {
 							for ( const context in strings[ file ] ) {
 								// Within the same file, sort translations by line.
-								const sortedTranslations = sortBy(
-									strings[ file ][ context ],
-									'comments.reference'
+								const sortedTranslations = sortByReference(
+									Object.values( strings[ file ][ context ] )
 								);
 
-								forEach(
-									sortedTranslations,
-									( translation ) => {
-										const { msgctxt = '', msgid } =
-											translation;
-										if (
-											! memo.hasOwnProperty( msgctxt )
-										) {
-											memo[ msgctxt ] = {};
-										}
-
-										// Merge references if translation already exists.
-										if (
-											isSameTranslation(
-												translation,
-												memo[ msgctxt ][ msgid ]
-											)
-										) {
-											translation.comments.reference = [
-												...new Set(
-													[
-														memo[ msgctxt ][ msgid ]
-															.comments.reference,
-														translation.comments
-															.reference,
-													]
-														.join( '\n' )
-														.split( '\n' )
-												),
-											].join( '\n' );
-										}
-
-										memo[ msgctxt ][ msgid ] = translation;
+								sortedTranslations.forEach( ( translation ) => {
+									const { msgctxt = '', msgid } = translation;
+									if ( ! memo.hasOwnProperty( msgctxt ) ) {
+										memo[ msgctxt ] = {};
 									}
-								);
+
+									// Merge references if translation already exists.
+									if (
+										isSameTranslation(
+											translation,
+											memo[ msgctxt ][ msgid ]
+										)
+									) {
+										translation.comments.reference = [
+											...new Set(
+												[
+													memo[ msgctxt ][ msgid ]
+														.comments.reference,
+													translation.comments
+														.reference,
+												]
+													.join( '\n' )
+													.split( '\n' )
+											),
+										].join( '\n' );
+									}
+
+									memo[ msgctxt ][ msgid ] = translation;
+								} );
 							}
 
 							return memo;