+++ packages/react-native-editor/__device-tests__/gutenberg-editor-lists-end.test.js
@@ -5,7 +5,7 @@ import { blockNames } from './pages/editor-page';
 import testData from './helpers/test-data';
 
 describe( 'Gutenberg Editor tests for List block (end)', () => {
-	it( 'should be able to end a List block', async () => {
+	it.skip( 'should be able to end a List block', async () => {
 		await editorPage.addNewBlock( blockNames.list );
 		const listBlockElement = await editorPage.getListBlockAtPosition();